New: Turns env files into etcd key / values
Added a new container which watches a folder of envrionment files.  Whenever
one is added or changed, it updates the /mayfly/environments keys to publicise
the new environment.

Known limitations include: no way to delete an environment; two files can duel
to control the configuration of an environment but remain in the folder
(whichever is updated most recently wins).  There may also be some race
conditions / unexpected behaviour because each of the keys describing a bit
of the environment is updated sequentially while haproxy will start updating
itself as soon as the first change is noted.